[aqi] Fix precision loss for low PM concentration values
Use floating-point math throughout the AQI calculation to prevent
integer truncation when PM values are below 1 µg/m³. Previously,
PM2.5 values like 1.8 would truncate to 1 before calculation,
and integer division would produce AQI values of 0-5 instead of
the correct 0-10 range for clean indoor air.

Changes:
- Accept float parameters in get_aqi() interface
- Use float arrays for PM concentration breakpoints
- Perform interpolation with float math, round only at final output

+++ esphome/components/aqi/aqi_sensor.cpp
@@ -44,8 +44,7 @@ void AQISensor::calculate_aqi_() {
     return;
   }
 
-  uint16_t aqi =
-      calculator->get_aqi(static_cast<uint16_t>(this->pm_2_5_value_), static_cast<uint16_t>(this->pm_10_0_value_));
+  uint16_t aqi = calculator->get_aqi(this->pm_2_5_value_, this->pm_10_0_value_);
   this->publish_state(aqi);
 }
